Added the two async functions in classes and created an authentication folder --anubrotoGhose
